# docs/swagger.yaml
definitions:
...
  project.InjectOperation:
    properties:
      archived:
        type: boolean
      customFilenames:
        items:
          type: string
        type: array
      customPayloads:
        items:
          type: string
        type: array
      doNotRecord:
        type: boolean
      error:
        type: string
      fuzzDB:
        items:
          type: string
        type: array
      guid:
        type: string
      host:
        type: string
      injectDescription:
        type: string
      iterateFrom:
        type: integer
      iterateTo:
        type: integer
      objectType:
        type: string
      percentCompleted:
        description: Parts of the object which cannot be set by JSON
        type: integer
      request:
        items:
          $ref: '#/definitions/project.InjectOperationRequestPart'
        type: array
      requestsMadeCount:
        type: integer
      ssl:
        type: boolean
      title:
        type: string
      totalRequestCount:
        type: integer
      url:
        type: string
    type: object
...
  proxy.InterceptSettings:
    properties:
      browserToServer:
        type: boolean
      serverToBrowser:
        type: boolean
    type: object
...
  proxy.PayloadEntry:
    properties:
      filename:
        type: string
      isDirectory:
        type: boolean
      resourcePath:
        type: string
      samplePayloads:
        items:
          type: string
        type: array
      subEntries:
        items:
          $ref: '#/definitions/proxy.PayloadEntry'
        type: array
      title:
        type: string
    type: object
...
host: localhost
info:
  contact:
    email: support@forensant.com
    name: API Support
    url: https://forensant.com/support
  description: This provides the common functions which are relied upon by the Proximity
    frontends.
  license:
    name: Commercial
    url: https://proximity.forensant.com/terms
  termsOfService: https://forensant.com/terms/
  title: Proximity Core
  version: "1.0"
paths:
...
  /inject_operations:
    get:
      description: gets a list of all injection operations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.InjectOperation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Inject Operations
      tags:
      - Injection Operations
  /inject_operations/{guid}/archive:
    patch:
      description: updates the the archived status of an inject operation
      parameters:
      - description: inject operation guid
        in: path
        name: guid
        required: true
        type: string
      - description: archive status to set
        in: formData
        name: archive
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Archive Inject Operation
      tags:
      - Injection Operations
  /inject_operations/{guid}/title:
    patch:
      description: updates the title of an inject operation
      parameters:
      - description: inject operation guid
        in: path
        name: guid
        required: true
        type: string
      - description: title to set
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Inject Operation Title
      tags:
      - Injection Operations
  /inject_operations/{path}:
    get:
      description: gets a single inject operation
      parameters:
      - description: The GUID of the request to fetch
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.InjectOperation'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Inject Operation
      tags:
      - Injection Operations
  /inject_operations/fuzzdb_payload:
    get:
      description: gets a specific fuzzdb file
      parameters:
      - description: The file path of the fuzzdb file to fetch the payload for
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a fuzzdb file
      tags:
      - Injection Operations
  /inject_operations/payloads:
    get:
      description: gets all available payloads available for injection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proxy.PayloadEntry'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets injection payloads
      tags:
      - Injection Operations
  /notifications:
    get:
      description: websocket endpoint to stream data as it is inserted/modified
      parameters:
      - description: JSON object (key:value) where the returned objects will be filtered
          by the values
        in: query
        name: objectfieldfilter
        type: string
      - description: additional filter to apply to the objects (behaviour is object
          dependent)
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Stream updates
      tags:
      - Misc
  /out_of_band/url:
    get:
      description: gets a unique URL which can be used to test out of band interactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Out of Band URL
      tags:
      - Out of Band
...
  /proxy/ca_certificate.pem:
    get:
      description: returns the certificate authority root certificate
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets the root CA
      tags:
      - Proxy
  /proxy/intercept_settings:
    get:
      description: get intercept settings
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proxy.InterceptSettings'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Intercept Settings
      tags:
      - Settings
    put:
      description: set intercept settings
      parameters:
      - description: Proxy Intercept Settings Object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/proxy.InterceptSettings'
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Intercept Settings
      tags:
      - Settings
  /proxy/intercepted_requests:
    get:
      description: gets a list of all requests which have been intercepted, which
        are awaiting a response
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.InterceptedRequest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Intercept Requests
      tags:
      - Proxy
  /proxy/set_intercepted_response:
    put:
      description: set how an intercepted request will be responded to
      parameters:
      - description: Proxy Intercept Response Object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/proxy.InterceptedRequestResponse'
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Modify Intercepted Request
      tags:
      - Proxy
  /proxy/settings:
    get:
      description: get proxy settings
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proxy.ProxySettings'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Proxy Settings
      tags:
      - Settings
    put:
      description: set proxy settings
      parameters:
      - description: Proxy Settings Object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/proxy.ProxySettings'
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Proxy Settings
      tags:
      - Settings
  /requests:
    get:
      description: gets a list of all requests
      parameters:
      - description: Scan ID, can be multiple separated by semi-colons
        in: query
        name: scanid
        type: string
      - description: Only show requests which contain the filter string in the url,
          request, response, etc
        in: query
        name: filter
        type: string
      - description: Only show requests which contain the given string in the URL
        in: query
        name: url_filter
        type: string
      - description: Column to sort by (default time)
        in: query
        name: sort_col
        type: string
      - description: Column direction to sort by (default asc)
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.Request'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Requests
      tags:
      - Requests
  /requests/{base_guid}/compare/{compare_guid}:
    get:
      description: compares two requests and returns the differences
      parameters:
      - description: Base Request guid
        in: path
        name: base_guid
        required: true
        type: string
      - description: Request to Compare guid
        in: path
        name: compare_guid
        required: true
        type: string
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.RequestDifference'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Compare Two Requests
      tags:
      - Requests
  /requests/{guid}:
    get:
      description: gets a specific request
      parameters:
      - description: The GUID of the request to fetch
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.RequestSummary'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get A Request
      tags:
      - Requests
  /requests/{guid}/contents:
    get:
      description: gets the full request and response of a given request
      parameters:
      - description: Request GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.RequestResponseContents'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Request and Response
      tags:
      - Requests
  /requests/{guid}/notes:
    patch:
      description: updates a specific request's notes
      parameters:
      - description: The GUID of the request to update
        in: path
        name: guid
        required: true
        type: string
      - description: The notes for the request
        in: body
        name: notes
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Request Notes
      tags:
      - Requests
  /requests/{guid}/partial_data:
    get:
      description: gets part of the request/response. will attempt to return at least
        5MB of data to cache
      parameters:
      - description: Request guid
        in: path
        name: guid
        required: true
        type: string
      - description: Offset to request from
        in: query
        name: from
        required: true
        type: integer
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.PartialRequestResponseData'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Request/Response Data
      tags:
      - Requests
  /requests/{guid}/payloads:
    patch:
      description: sets the payloads associated with a specific request
      parameters:
      - description: The GUID of the request to update
        in: path
        name: guid
        required: true
        type: string
      - description: A JSON Object containing the payloads in {'key':'value'} format
        in: body
        name: payloads
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Request Payloads
      tags:
      - Requests
  /requests/make:
    post:
      consumes:
      - application/json
      description: makes a single request to a given server
      parameters:
      - description: Make Request Parameters in JSON format
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/proxy.MakeRequestParameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Make a single request
      tags:
      - Requests
  /requests/queue:
    post:
      description: add a request to the queue for scanning sites
      parameters:
      - description: Request Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/proxy.AddRequestToQueueParameters'
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add Request to Queue
      tags:
      - Requests
  /requests/sitemap:
    get:
      description: gets a list of all paths observed by the proxy
      parameters:
      - description: An optional filter on the query to restrict to specific paths
        in: query
        name: parent
        required: true
        type: string
      - description: An optional filter on the query to restrict to the paths to those
          seen for a particular scan
        in: query
        name: scan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets the sitemap
      tags:
      - Requests
  /script_groups:
    get:
      description: gets a list of all script groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.ScriptGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Script Groups
      tags:
      - Scripting
    post:
      description: adds or updates a script group
      parameters:
      - description: Script Group details in JSON format
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/project.ScriptGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add/Update Script Group
      tags:
      - Scripting
  /script_groups/{guid}:
    get:
      description: gets a specific script group
      parameters:
      - description: Script group guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ScriptGroup'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Script Group
      tags:
      - Scripting
  /script_groups/{guid}/archive:
    patch:
      description: updates the archived status of a script group
      parameters:
      - description: script group guid
        in: path
        name: guid
        required: true
        type: string
      - description: archive status to set
        in: formData
        name: archive
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Archive Script Group
      tags:
      - Scripting
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"